requirements.txt
streamlit
pandas
numpy
opencv-python
deepface
Pillow
fpdf
tf-keras
